requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
common==0.1.2
contourpy==1.2.0
cycler==0.12.1
fairscale==0.4.13
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
idna==3.6
iopath==0.1.10
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
plotly==5.18.0
portalocker==2.8.2
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tk==0.1.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.0
yacs==0.1.8
